eas_deployments/requirements.txt
connexion[swagger-ui]
connexion[uvicorn]
Flask-PyMongo
flask-bcrypt
flask-cors
flask-jwt-extended
mongoengine
pymongo[srv]